Update HTML/CSS/JS
Expanded the capabilities of the hamburger menu. It is now correctly invoked from all paged of the website, and I also added a scrolling function so it hides when scrolling down and reveals itself when scrolling up. The power of java is now coursing through my veins, and it will only grow stronger as I keep learning >=)

Besides all that, I have also started to adapt the design of the sections to better fit the mobile layout. Still incomplete and with lots of placeholders though.