Final Delivery Commit
With the deadline approaching, there was no time to try and organize the workflow. Instead I opted to go into a mad sprint to the finish line and I achieved what I set out to. 

Now I still have about a day left for the final deadline, with a few other tasks to complete. In the event I finish everything else early (unlikely), I might come back to try and add that extra bit of polish here and there. Otherwise, this may be considered the final product -even though it doesn't show the full extent of my capabilities.